+++ NEWS.md
@@ -4,6 +4,7 @@
 
 * Adding possibility to add internal libraries through the GUI (#159)
 * Expose more parameters to the GUI (#159)
+* Fix adducts and remove nitrogen rule
 * Updated benchmarking steps
 
 # timaR 2.9.6